Revert "Issue 52 sign up screen"
