+++ MusicMind/MusicMind/SignUp/MobileViewController.swift
@@ -1,0 +1,32 @@
+//
+//  MobileViewController.swift
+//  MusicMind
+//
+//  Created by Angel Contreras on 3/30/17.
+//  Copyright © 2017 MusicMind. All rights reserved.
+//
+
+import UIKit
+
+class MobileViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    var newUser: User?
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
