+++ MusicMind/MusicMind/SignUp/NameViewController.swift
@@ -14,18 +14,12 @@ class NameViewController: UIViewController {
     @IBOutlet weak var lastNameTextField: UITextField!
     @IBOutlet weak var signUpButton: UIButton!
     
-    var newUser = User()
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         self.firstNameTextField.delegate = self
         self.lastNameTextField.delegate = self
         // Do any additional setup after loading the view.
-        
-        //DELETE
-        firstNameTextField.text = "Angel"
-        lastNameTextField.text = "Contreras"
     }
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -38,8 +32,16 @@ class NameViewController: UIViewController {
                 return
         }
         
-        newUser.firstName = firstName
-        newUser.lastName = lastName
+        let appDelegate = UIApplication.shared.delegate as! AppDelegate
+        let user = appDelegate.user
+//        user.firstName = firstName
+//        user.lastName = lastName
+        
+        if segue.identifier == "toBday",
+            let birthdayViewController = segue.destination as? BirthdayViewController {
+            
+//            birthdayViewController.newUser = user
+        }
     }
     
 }
