+++ MusicMind/MusicMind/UserLoginCredentials.swift
@@ -25,14 +25,12 @@ class UserLoginCredentials: NSObject {
         super.init()
     }
 
-
     var firebaseUserEmail: String? {
         get {
             
             return keychain.get(KeychainKey.firebaseUserEmail.rawValue)
         }
         set {            
-
             if let unwrappedNewValue = newValue {
                 keychain.set(unwrappedNewValue, forKey: KeychainKey.firebaseUserEmail.rawValue)
             } else {
@@ -54,4 +52,5 @@ class UserLoginCredentials: NSObject {
             }
         }
     }
+
 }
